requirements.txt
# This is a requirement.txt file
# to install all these requirements:
# `pip install -r requirements.txt`
# Make sure you have pip installed!
# To update just use `pip install -r requirements.txt --upgrade`
torch
torchvision
Pillow                # PIL
scikit-image          #skimage
tqdm
opencv-python         # cv2
trimesh
PyOpenGL
ffmpeg
matplotlib
pycocotools
rembg
